+++ src/features/create-new-password/ui/CreateNewPasswordForm.stories.tsx
@@ -33,6 +33,6 @@ export const Default: Story = {
     await userEvent.type(canvas.getByLabelText('Password confirmation'), '654321')
     await userEvent.click(submitButton)
 
-    await expect(canvas.getByText('Passwords must match')).toBeInTheDocument()
+    await expect(canvas.getByText('The passwords must match')).toBeInTheDocument()
   },
 }